+++ src/TransitConnection.java
@@ -1,0 +1,14 @@
+import java.util.Queue;
+
+public class TransitConnection {
+    public TransitConnection( String id, float distance, float time ){
+        id = this.idOfConnectedNode;
+        distance = this.distance;
+        time = this.time;
+    }
+
+    Queue<Commuter> transitConnection;
+    String idOfConnectedNode;
+    float distance;
+    int time;
+}
